# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ludus UX API
  description: |
    REST API for the Ludus UX web application — a Next.js frontend for managing
    [Ludus](https://ludus.cloud) cyber range infrastructure, including VM ranges,
    content sources (sync / force re-install), templates, user management,
    VNC/SPICE console access, GOAD (Game of Active Directory) lab orchestration,
    and LudusHound BloodHound→Ludus deploys.

    ## Authentication

    All protected endpoints require a valid session cookie set by `POST /api/auth/login`.
    The cookie is AES-256-GCM encrypted and HTTP-only. Sessions expire after 24 hours.

    **Admin impersonation headers** (admin sessions only):
    - `X-Impersonate-As` — target username
    - `X-Impersonate-Apikey` — target user's Ludus API key
    - `X-Ludus-Admin: true` — route to admin API port (8081)
    - `X-Ludus-User` — override Ludus user context
  # Keep in sync with package.json / APP_VERSION (src/lib/changelog.ts).
  version: 1.4.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Ludus UX application server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ludus_session
      description: AES-256-GCM encrypted HTTP-only session cookie set by the login endpoint.
...
  schemas:
...
    RuntimeSettings:
      type: object
      properties:
        ludusUrl:
          type: string
          format: uri
          example: https://192.168.1.10:8080
        ludusAdminUrl:
          type: string
          format: uri
          example: https://192.168.1.10:8081
        sshHost:
          type: string
        sshPort:
          type: integer
          default: 22
        goadPath:
          type: string
          example: /opt/GOAD
        ludusInstallPath:
          type: string
          example: /opt/ludus
        rootApiKey:
          type: string
        proxmoxSshUser:
          type: string
          example: root
        proxmoxSshPassword:
          type: string
          format: password

    # ---- Range / VM ----
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: No valid session cookie or session has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Authenticated but insufficient privileges (admin required).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get runtime settings
      description: Returns the current runtime configuration (Ludus URL, SSH config, GOAD path, etc.).
      operationId: getSettings
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Settings]
      summary: Update runtime settings
      description: Persists updated runtime settings to SQLite. Admin required.
      operationId: updateSettings
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ludusUrl:
                  type: string
                  format: uri
                ludusAdminUrl:
                  type: string
                  format: uri
                sshHost:
                  type: string
                sshPort:
                  type: integer
                goadPath:
                  type: string
                ludusInstallPath:
                  type: string
                proxmoxSshUser:
                  type: string
                proxmoxSshPassword:
                  type: string
                  format: password
      responses:
        '200':
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/settings/install-lux-host:
    post:
      tags: [Settings]
      summary: Install the scoped lux-host sudo rule
      description: |
        Writes `/usr/local/sbin/lux-host` and a sudoers rule for that helper only.
        `PROXMOX_SSH_USER` does not have to be root. A non-root account is valid when
        it can run `sudo -n /usr/local/sbin/lux-host`. If that account is not in
        sudoers. `rootPassword` is required. LUX SSHes as root to write the files and does not save it.
        Admin required.
      operationId: installLuxHost
      security:
        - sessionCookie: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                sshHost:
                  type: string
                sshPort:
                  type: integer
                proxmoxSshUser:
                  type: string
                proxmoxSshPassword:
                  type: string
                  format: password
                rootPassword:
                  type: string
                  format: password
                  description: Used once to SSH as root. Not stored.
      responses:
        '200':
          description: The SSH account can run the helper without a password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
                required: [ok, message]
        '400':
          description: The SSH user is invalid, a password contains a newline, or the account cannot install the helper.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          description: The helper was written but verification did not return uid 0, or the install command failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --------------------------------------------------------------------------
  # Releases
  # --------------------------------------------------------------------------
...